+++ components/Explore/SwipeCard/SwipeCard.tsx
@@ -1,4 +1,5 @@
 import { PanInfo, motion, useMotionValue, useTransform } from "framer-motion";
+import { useRef } from "react";
 
 const DRAG_LIMIT = 300;
 
@@ -23,6 +24,7 @@ export const SwipeCard: React.FC<ISwipeCardProps> = ({
   isLike,
   setIsLike,
 }) => {
+  const constraintsRef = useRef(null);
   const x = useMotionValue(0);
   const rotate = useTransform(x, [-DRAG_LIMIT, DRAG_LIMIT], [-22.5, 22.5]);
   const background = useTransform(
@@ -55,14 +57,15 @@ export const SwipeCard: React.FC<ISwipeCardProps> = ({
 
   return (
     <motion.div
+      ref={constraintsRef}
       style={{ background }}
       className="w-full h-[calc(100svh-4rem)] sm:min-h-screen flex justify-center items-center fixed"
       exit={{ opacity: 0, transition: { duration: 0.4 } }}
     >
       <motion.div
         drag
-        dragConstraints={{ top: 0, left: 0, right: 0, bottom: 0 }}
-        dragElastic={0.1}
+        dragConstraints={{ left: 0, right: 0, top: 0, bottom: 0 }}
+        dragElastic={0.3}
         onDragEnd={dragEndHandler}
         onDrag={dragHandler}
         animate={{ scale: 1, opacity: 1 }}
