+++ pages/likes/index.tsx
@@ -3,10 +3,10 @@ import { getServerSession } from "next-auth";
 import { authOptions } from "../api/auth/[...nextauth]";
 import { prisma } from "@/lib/prisma";
 import type { Recipe } from "@/interfaces";
-import { Pagination } from "@/components/Likes/Pagination/Pagination";
-import { LikedRecipe } from "@/components/Likes/LikedRecipe/LikedRecipe";
-import { Title } from "@/components/ui/Title/Title";
-import { ContentWrapper } from "@/components/ui/ContentWrapper/ContentWrapper";
+import { Pagination } from "@/components/Likes/Pagination";
+import { LikedRecipe } from "@/components/Likes/LikedRecipe";
+import { Title } from "@/components/ui/Title";
+import { ContentWrapper } from "@/components/ui/ContentWrapper";
 
 interface ILikesPageProps {
   recipes: Recipe[];
